GS-1655 Little icons on choose city. Limit the number of cities displayed.

+++ src/test/gs/web/geo/GeoControllerTest.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2005 GreatSchools.net. All Rights Reserved.
- * $Id: GeoControllerTest.java,v 1.5 2006/02/27 18:16:13 apeterson Exp $
+ * $Id: GeoControllerTest.java,v 1.6 2006/03/20 23:20:37 apeterson Exp $
  */
 
 package gs.web.geo;
@@ -47,7 +47,7 @@ public class GeoControllerTest extends BaseControllerTestCase {
         assertNotNull(modelAndView.getModel().get("lon"));
 
         assertNotNull(modelAndView.getModel().get("scale"));
-        assertEquals(new Integer(7), modelAndView.getModel().get("scale"));
+        assertEquals(new Integer(5), modelAndView.getModel().get("scale"));
     }
 
     public void testAlameda() throws Exception {
